Refactor to isolate controller from model.
Previously it had been expected that the model would invoke methods on the
controller.  This removes one instance of this incorrect assumption (the
execute_webhook_hash method on the model).  Prefering instead for the API layer
to get the information about the webhook from the models and use that to
get the group and policy information necessary to pass to the controller.