scheduler consistency is QUORUM
It is possible that 1 api node deletes event and before it gets propogated to other cass nodes, other
api node picks the same event causing the same policy to get executed twice. So, read and writing at
quorum that guarantees consistency. The slower performance due to this is not really a concern since it
is happening in background.