Add CassAdmin, MockAdmin, IAdmin. No more globals.
CassAdmin and MockAdmin both implement IAdmin. IAdmin defines the
interface for admin methods, and currenly only contains `get_metrics`.

get_store and set_store are not used in the OtterAdmin app. The store
is passed as the first argument to OtterAdmin, and defaults to MockAdmin
if none is passed in.