Fix tests for API Options, Add tests for admin port
Tests for 'port' in the APIOptions were failing because
`assert_called_with` was being used and a new subservice was added. This
made calls to `Site` happen twice. `assert_called_with` only looks at
the most recent call, and since the admin subservice was added after the
root service, the test was failing with the wrong call.