adding user agent so the things we call with HTTP know it is us.
Looking at our code, it looks like we're not setting many HTTP header variables.

Most specifically, we're not sending the User-Agent.

I was talking with the CLB team and I talked them out of the idea of doing
special purpose behaviors for us -- that's kinda a violation of the basic
principles of distributed systems.  But they still would like to be able to
do filtering for their logs so they can better track our errors on both sides.