partitionining in subprocess; NO TESTS!
Scheduler delegates the partitoning to sub-process and provides reset()
that will reset the scheduler with new path in new process. This way
partitioning can be fixed with a new path when it breaks. This method is
exposed with /scheduler_reset URI

- new file partition.py that does the partitoning and outputs the
partition on stdout on regular basis
- this is called with necessary command-line args from scheduler
- scheduler is not TimerService anymore and relies on subproc to give
buckets on timely manner
- reset() starts new partition in new process and leaves old process
running for debugging purpose