Extract converge function from duplicates
This is a refactor of the maybe_execute_scaling_policy and
obey_config_change into a new function, converge, which is intended
to be a central point which we can extend to support Heat-based
orchestration.

Some minor changes in the wording of log messages have been made,
since the code duplicates did logging differently.