Refactor integration tests away from unit tests.
This is a moderate-sized PR, but I couldn't really break it up into
smaller pieces without it making sense.  It refactors the work invested
in otter/test/integration/lib.py into component modules which can me
"from otter.integrat.lib import"-ed, making it easier to manage in the
future as the feature set grows.  Work on lib.py has already been
reviewed previously, so despite the apparent size of the PR, relatively
little needs to be reviewed.  In particular, the reviewer should focus
on the refactoring action, and not so much on the code.

The actual "new code" in this PR sits in the new tests/ subdirectory --
the test_scaling.py module.  This module isn't finished yet; however,
refactoring was a good idea before continuing for two reasons:

1) Refactoring removes lib.py and test_lib.py from the "make unit"
execution path.

2) Refactoring allows me to commit my integration test work to the
repository in conjunction with its co-developed support library.
Previously, this work contained credentials hard-wired into the code.

Fixes #1074