switch to {tenant}_{group}_{random} for the format of the dirty flag
this eliminates the use of version numbers on dirty flags, replacing
them with the use of unique flags for each request for convergence.

This isn't as efficient as versions, since it can lead to extra
unnecessary convergence iterations when multiple convergence
requests happen rapidly.

However, the benefit of this format is that we can use a ChildrenWatch
on the /groups/divergent directory and get notified of all new dirty
flags so we can immediately start convergence without delay.