WIP: First convergence-specific test I'm aware of.
This test aims to exercise the case where a user deletes a Nova server
from a scaling group out-of-band.

It's not finished yet, but it runs through to completion and doesn't
mess anything up.

It works by creating a scaling group with N_SERVERS (constant in the
source code).  It then deletes half of those servers through Nova's API.
What comes next is a subject for a future PR.