RST conversion cleanup
Restructured directory so that conf.py is in same directory as the RST
source files.

Changed index.rst to include on level 1 header files.
Update level 1 header files (general-api.information.rst, for example)
to include files with level 2 and lower headings.

Changed extension on included files to *.txt to get correct nesting in
table of contents.