buckets_acquired different return
zkpartitioner changed to respect deferred returned from got_buckets
callback which in convergerservice will be one that gets fired when all
convergence is completed. We don't need that since convergence is
idempotent. it is ok to stop otter in middle of convergence cycles which
is why the deferred is returned in a tuple as it will be ignored