better logging when triggering convergence
not binding log inside trigger_convergence. instead expecting callers to
do as they will have tenant, group and any other kind of trans id.

trigger_convergence does not skip error. instead it logs *and* raises
it. This way it will propogate up to REST resulting in 500 which is
better than 200 but not doing anything