Fix throttling so that it actually works
Previously, the lock object used to implement throttling in cloud_client
was not shared between calls, so in essence there was absolutely no
useful locking going on.

This change uses a WeakLocks object to keep track of the locks so
that they're shared between calls. (this is a bit overkill
for the current situation, since we have a fixed number of locks
so we could just put them in a global data structure, but with
upcoming changes to make locking be per-tenant, WeakLocks makes more
sense).