health monitor info in new CLB object
It was generally incorrect to put health monitor in CLBDescription as
it is supposed to represent Node-CLB mapping - NOT a real CLB. This
became apparent when I realized that CLBDescription is created from launch config.
Moreover health monitor wouldn't be known if there are no nodes in CLB
which is bad.

Hence, created simple CLB object that only contains health monitor. It
is passed to planner as dict of opaque objects not directly used
by the planner. This object is got from `get_clb_contents` as part of
gathering step and passed on to planner.

In planner, LB object is passed to `add_server_to_lb` and
`change_lb_node` for specific LB providers to use if needed. Later on
it can be passed on to other parts of planner if need arises.