Updated pyOpenSSL to 16.2.0 in requirements.txt.
You can't activate the ve and get 'pip list' to work if you don't do this. I
don't understand all of the details, but the version of pyOpenSSL from before
this change has aged enough to really break things. This has to get updated for
anything approaching a working build.