Work-around probable mimic bug by randomizing 'trial' test run order in docker-compose.yml
I added -z 1 to the 'trial' command in the docker-compose.yml to randomize the
order (with a seed of positive integer 1). In this case, if you run "trial" by
itself it tends to freeze up mimic pretty hard. It works fine running it along
side the 'cafe' tests and just letting everything go with 'docker-compose up'.
It also works fine if you randomize the tests.

Admittedly, I had to raise pyOpenSSL version to 16.2.0 from 16.0.0 and
crytography to 1.9 from 1.4 to get the Dockerfile to work. Also, I use a version
of mimic 246 commits behind, to match the version we originally had here.

I think this probably means we have some mimic bug, but we lost the last mimic
developer sometime around July, 2018, and I haven't tried a newer version yet.
This will work for now.