Try to pull mimic from an alternative repo so that we can update it.
In this case, we don't have rights to commit to rackerlabs/mimic. In fact, I
don't think I can even hit the "fork" button on it, although I might've just
encountered a problem with forking to the same org (rackerlabs) because it has
the same name. I don't remember for sure, but we'll want the repo in rackerlabs
anyway, so I cloned it locally and made a differently-named repo in rackerlabs
for it. We can commit here. Theoretical syncing with upstream will just have to
go through a local git repo. Upstream has dead development, though, probably not
an issue. The sync would involve:

* a clone of mimic or autoscale-mimic locally
* a remote to the original repo, probably called "origin"
* a remote to autoscale-mimic repo, probably called "autoscale-mimic"
* fetch/pull/merge from origin to local repo
* push to autoscale-mimic

Basically, a fork without the benefit of GitHub's "fork" button.

To start with, I have created branch "autoscale" from commit 594dd5c, which
came from otter/requirements/mimic.txt, then I check that out in
requirements.txt

In this case, we had rackerlabs/mimic as public, and I have
rackerlabs/autoscale-mimic as A private repo, so to get this to work locally, I
had to use ssh as the protocol in requirements.txt, and add a stanza like this
to my ~/.ssh/config:

Host github.com
    User git
    IdentityFile ~/.ssh/id_rsa

since my key gets me access. This sufficed for local testing.

Ultimately, I needed to do the same thing on the jenkins server with the SSH
config, with the "deploy" user key (id_rsa(.pub)?). However, additionally, I
needed to add that key as a *deploy key* for the autoscale-mimic repo:

https://developer.github.com/v3/guides/managing-deploy-keys/#deploy-keys

With a fake full login as the jenkins user, I confirmed this command worked on
the jenkins server (and, of course, tried to run the integration tests, which
pulls/installs mimics and deploys to the ci server):

git clone -q ssh://github.com/rackerlabs/autoscale-mimic.git

This used the SSH key the jenkins server had, which we had added as a deploy
key to the autoscale-mimic repo. GitHub appears to always take the username
"git" for these kinds of things.

Furthermore, I confirmed the site-cookbooks/mimic cookbook tries to install the
mimic it finds in otter/requirements/mimic.txt. This should need no other
changes; mimic always seems to come from the source specified in
otter/requirements/mimic.txt